+++ src/main/java/org/symphonyoss/integration/webhook/salesforce/parser/v1/CommonSalesforceParser.java
@@ -4,11 +4,14 @@ import static org.symphonyoss.integration.messageml.MessageMLFormatConstants.MES
 import static org.symphonyoss.integration.messageml.MessageMLFormatConstants.MESSAGEML_START;
 import static org.symphonyoss.integration.parser.ParserUtils.presentationFormat;
 import static org.symphonyoss.integration.parser.SafeString.EMPTY_SAFE_STRING;
+import static org.symphonyoss.integration.webhook.salesforce.SalesforceConstants.EMAIL_ADDRESS;
+import static org.symphonyoss.integration.webhook.salesforce.SalesforceConstants.INTEGRATION_NAME;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.symphonyoss.integration.entity.Entity;
+import org.symphonyoss.integration.entity.EntityBuilder;
 import org.symphonyoss.integration.entity.model.User;
 import org.symphonyoss.integration.messageml.MessageMLFormatConstants;
 import org.symphonyoss.integration.model.message.Message;
@@ -22,6 +25,7 @@ import org.symphonyoss.integration.webhook.salesforce.parser.SalesforceParser;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
@@ -54,6 +58,46 @@ public class CommonSalesforceParser implements SalesforceParser {
 
   private String salesforceUser;
 
+  /**
+   * Searches a matching Symphony user for the given Salesforce user email.
+   * @param email email to be found
+   * @return user information, including a matching Symphony ID (if any)
+   */
+  private User getUser(String email) {
+    return userService.getUserByEmail(salesforceUser, email);
+  }
+
+  protected void createMentionTagFor(Entity mainEntity, String userEntityName) {
+    Entity userEntity = mainEntity.getEntityByName(userEntityName);
+    if(userEntity == null){
+      return;
+    }
+    String email = getEmail(userEntity);
+    if (!StringUtils.isEmpty(email)) {
+      User user = getUser(email);
+      EntityBuilder.forEntity(userEntity).nestedEntity(user.getMentionEntity(INTEGRATION_NAME));
+    }
+  }
+
+  /**
+   * Creates the mention tag for the entity received.
+   * @param mainEntity Main entity to find the nested entity
+   * @param nestedEntityName Entity to find in the main entity
+   * @param userEntityName Entity name for set the mention tag
+   */
+  protected void createListOfMentionsFor(Entity mainEntity, String nestedEntityName, String userEntityName){
+    Entity entity = mainEntity.getEntityByType(nestedEntityName);
+    if (entity != null) {
+      for (Entity nestedEntity : entity.getEntities()) {
+        createMentionTagFor(nestedEntity, userEntityName);
+      }
+    }
+  }
+
+  private String getEmail(Entity entity) {
+    return entity.getAttributeValue(EMAIL_ADDRESS);
+  }
+
   @Override
   public String parse(Entity entity) throws SalesforceParseException {
     return null;
@@ -66,7 +110,7 @@ public class CommonSalesforceParser implements SalesforceParser {
 
   @Override
   public List<String> getEvents() {
-    return null;
+    return Collections.emptyList();
   }
 
   @Override
@@ -300,5 +344,4 @@ public class CommonSalesforceParser implements SalesforceParser {
 
     return presentationFormat(FORMATTED_STRING, updatedFields);
   }
-
 }
